soc/intel/common/block/cse/cse.c: fix incorrect cse_get_me_disable_mode ret value
Fix incorrect return value when CONFIG_DRIVERS_EFI_VARIABLE_STORE is
not selected.

+++ src/soc/intel/common/block/cse/cse.c
@@ -1099,7 +1099,7 @@ uint8_t cse_get_me_disable_mode(void)
 
 	return var;
 #else
-	return UINT_MAX;
+	return UINT8_MAX;
 #endif
 }
